Added a few algorithm options on missing parents, default option functionally equivalent as before
The SRD sections are not contiguous, the previous commit fixes this by making
subsections with different font sizes share the same toc level.

In addition to that "rebase", this commit introduces two new methods to fix this
- "insert" inserts the necessary empty sections
- "remove" removes the section completely
since PyMuPDF errors out when the first toc entry is not level 1 or any entry is
not contiguous with the previous entry.

Interestingly, the newly released Spanish-localized SRD at
    https://www.dndbeyond.com/community-update
eg
    https://media.wizards.com/2023/downloads/dnd/SRD_CC_v5.1_ES.pdf

does have a TOC, but it's doesn't reflect the document accurately (eg the top
level section "Race", etc is missing for all the sections, so first level of the
TOC is actually the subsections).

It does feel like the SRD is missing a "Classes" top level section, and the
class names a font size demotion, to mimic the "Races" section wrt the race
names

Files:

srdine.py
- Added "rebase", "insert" and "remove" options.
- Removed some commented out code leftovers

README.md
- Removed done TODO

Tests

- Viewed the generated document with Sumatra PDF, the document with the "rebase"
  method looks identical as before this commit and they are the same size and
  almost binary equal.
- Generated using the different missing section methods
- Compared "rebase" TOC against Spanish SRD which does have a TOC (they are
  different, this one is more accurate).